Won't override pointers leading into memory leaks
For some objects from JS to C struct, there are some pointers to arrays
or other structs, now they are freed before and then override by the
new object.
Uses C malloc and free instead of C++ allocations.
On another cases, will only write the available objects, instead of
deleting and recreating the objects. That could cause a problem to the
FLAC API.