refactor: use static/dynamic linked libFLAC
This changes how libFLAC is loaded and now no longer uses dynamic
library loading to use the bindings. Compilation will have a default
download-and-compile itself for FLAC, or externally provided FLAC.

The prebuild package will use the static linking for an easy to use, but
it will not have enabled the Ogg support. With some environment
variables, the installation can change to use an external provided
library or enable Ogg support with the internal compilation.