refactor: use Reference and call detach on buffers
To be able to use shared pointers from flac API to node.js, already used
buffers must be detached so they can be attached to a new buffer again
(if required). This behaviour is due to a change on Buffer by using
BackingStore from V8 (required by them). Two BackingStores cannot point
to the same memory address, so in order to reuse pointers, the buffers
must be detached after use.
The use of References is to simplify and avoid unecessary detach/attach
calls. Also allows to give some extra time to JS code to use the Buffer
before it is detached (but this is undeterministic, it can be detached
any time).
See related issue #25