refac: move audio manipulation to C++
In StreamDecoder and FileDecoder, there is an audio conversion process
that runs all time. Time spent in this operation is not as huge as
decoding process itself, but it's always nice to improve performance. By
moving this conversion to C++, the performance gain is almost x2 (not in
decoding process overall, just in this conversion), so for large
decoding processes, can reduce the time for some seconds. On a small
test in local it reduced the process by 2s (for a file 192KHz, 2
channels, 24 bit around 6 minutes).
For StreamEncoder and FileEncoder, depending on the input PCM data and
expected format, it can run some conversion too (which is different from
the decoder one). This process was already quick, but by moving it to
C++ it improved around x1.5 (x2 in some cases). Test using the same
tests, which some of them uses this conversion.
An overall, running the whole test suite in release mode reduced the
duration around 1s.