feat: decoders emit format + encoders understand format
Now the StreamDecoder and FileDecoder will emit `format` event when the
`STREAMINFO` metadata block is read from the flac decoder. This event
can then be used for any other transform or write stream that
understands the event, like the StreamEncoder or FileEncoder, as well as
other packages like `speaker`, or the event from other packages (like in
`wav`) can be read to configure the encoder.

NOTE: Possible breaking change in encoder: now by default `inputAs32` is
always false.